requirements.txt
asgiref==3.5.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-import-export==2.7.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.0
xlrd==2.0.1
xlwt==1.3.0
aniso8601==7.0.0
asgiref==3.5.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-import-export==2.7.1
et-xmlfile==1.1.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
promise==2.3
pytz==2021.3
PyYAML==6.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.0
text-unidecode==1.3
xlrd==2.0.1
xlwt==1.3.0
